dataprocessing2 diviser reste plus qu a norminer
